requirements.txt
google-cloud-translate==2.0.0
google-cloud-vision